% Generated by build script
\name{exo_discoveries}
\alias{exo_discoveries}
\title{Get discovery statistics by year and method}
\usage{
exo_discoveries()
}
\arguments{
\item{top}{Number of years to return (default 30)}
}
\value{tibble: disc_year, discoverymethod, count}
\description{Get discovery statistics by year and method}
