% Generated by build script
\name{exo_habitable}
\alias{exo_habitable}
\title{Find potentially habitable exoplanets}
\usage{
exo_habitable()
}
\arguments{
\item{top}{Number of results (default 50)}
}
\value{tibble: pl_name, hostname, disc_year, mass_earth, radius_earth,}
\description{Find potentially habitable exoplanets}
