% Generated by build script
\name{exo_planets}
\alias{exo_planets}
\title{Query confirmed exoplanets}
\usage{
exo_planets()
}
\arguments{
\item{where}{SQL WHERE clause (optional, e.g. "disc_year>2020")}
\item{top}{Number of results (default 50)}
}
\value{tibble: pl_name, hostname, disc_year, mass_earth, radius_earth,}
\description{Query confirmed exoplanets}
